Handle relative links in Markdown files
Particularly if the hosting application is Konqueror, all links
requested or emitted in signals by the part are expected to be
absolute URLs.  There will be various asserts if they are not.

Set the base URL of the QTextDocument, and use this to normalise
relative links before passing them on.