Support new look-up pattern of ui.rc files with KXmlGui6
qrc file do not support build-time logic by themselves.
Generating qrc files from templates would place them into the build dir,
but they require the files they reference to be in the same dir or subdirs.

So the most simple approach for now seems to be to just have separate full
copies in the sources, assuming they do not see much changes in the near
future and thus have only little risk of getting out-of-sync.